backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pypdf==4.0.1
pdfplumber==0.10.3
reportlab==4.0.8
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.9.0
python-dotenv==1.0.0
python-docx==1.1.0
